chore: minor improvements
create transactions for discount
when softshop is running a discount
after all the original transactions has been logged
debit transactions should also be logged for the discount
if a delivery discount of 50% is given, and original delivery fee is 1000, users are expected to pay 500 but the order model will still have the original 1000 in it.
Logistics partners and ledger will be credited 1000 and another debit transaction of 500 will be made to Ledger.